+++ src/main/java/plugily/projects/murdermystery/handlers/lastwords/LastWordsManager.java
@@ -1,3 +1,20 @@
+/*
+ * MurderMystery - Find the murderer, kill him and survive!
+ * Copyright (C) 2020  Plugily Projects - maintained by Tigerpanzer_02, 2Wild4You and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package plugily.projects.murdermystery.handlers.lastwords;
 
 import org.bukkit.configuration.ConfigurationSection;
@@ -9,10 +26,14 @@ import plugily.projects.murdermystery.utils.Debugger;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.Collectors;
 
+/**
+ * @author 2Wild4You & Tigerpanzer_02
+ * <p>
+ * Created at 19.02.2021
+ */
 public class LastWordsManager {
 
   private final ArrayList<LastWord> registeredLastWords = new ArrayList<>();